BE: [refactor] DTO 디렉터리 구조 분리 (request/response) #6
