+++ Backend/src/main/java/gradude/springVision/domain/user/service/AuthCommandService.java
@@ -1,6 +1,10 @@
 package gradude.springVision.domain.user.service;
 
-import gradude.springVision.domain.user.dto.*;
+import gradude.springVision.domain.user.dto.request.SignupRequestDTO;
+import gradude.springVision.domain.user.dto.request.TokenRequestDTO;
+import gradude.springVision.domain.user.dto.response.KakaoUserInfoResponseDTO;
+import gradude.springVision.domain.user.dto.response.LoginResponseDTO;
+import gradude.springVision.domain.user.dto.response.TokenResponseDTO;
 import gradude.springVision.domain.user.entity.User;
 import gradude.springVision.domain.user.repository.UserRepository;
 import gradude.springVision.global.auth.TokenProvider;
